# swagger/docApiDuvidas.yaml
openapi: 3.0.1
info:
  title: Forum do programa de bolsas.
  description: >-
    API para postagem de dúvidas e respectivas soluções, dentro do programa de
    bolsas da empresa Compasso/UOL.
  version: 1.0.0
  termsOfService: https://compassouol.com
  contact:
    name: Suporte a desenvolvedores
    email: compasso@hotmail.com
  license:
    name: 'Licença: GPLv3'
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:8080/
    description: API de teste
paths:
...
  /cursos:
    get:
      tags:
        - cursos
      summary: Recupera uma lista de cursos.
      description: Pode recuperar todos os cursos cadastradas
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursosDTO'
    post:
      tags:
        - cursos
      summary: Cadastra um novo curso
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CursoFormDTO'
      responses:
        '201':
          description: Cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursoDTO'
  /cursos/{id}:
    get:
      tags:
        - cursos
      summary: Recupera um curso de determinado ID
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursoDTO'
        '404':
          description: Curso não encontrado
    put:
      tags:
        - cursos
      summary: Atualiza um curso de determinado ID
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CursoFormDTO'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursoDTO'
        '404':
          description: Curso não encontrado
    delete:
      tags:
        - cursos
      summary: Remove o curso de determinado ID
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
        '404':
          description: Curso não encontrado
  /cursos/{id}/topicos:
    post:
      tags:
        - cursos
      summary: Adiciona um topico a um determinado curso
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopicoFormDTO'
      responses:
        '201':
          description: Cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicoDTO'
        '404':
          description: Curso ou topico não encontrados
    get:
      tags:
        - cursos
      summary: Recupera todos tópicos de determinado curso
      description: Pode recuperar todos os tópicos de determinado curso pelo ID
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
        - name: titulo
          in: query
          schema:
            type: string
          required: false
        - name: curso
          in: query
          schema:
            type: string
          required: false
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursosDTO'
  /cursos/{id1}/topicos/{id2}:
    get:
      tags:
        - cursos
      summary: Retorna um topico de determinado curso
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicoDTO'
        '404':
          description: Curso ou topico não encontrados
    put:
      tags:
        - cursos
      summary: Atualiza um topico de determinado curso
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopicoFormDTO'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicoDTO'
        '404':
          description: Curso ou topico não encontrados
    delete:
      tags:
        - cursos
      summary: Remove um determinado topico de um determinado curso
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
        '404':
          description: Curso ou topico não encontrados
  /cursos/{id1}/topicos/{id2}/fechar:
    patch:
      tags:
        - cursos
      summary: Fecha um determinado tópico de determinado curso
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicoDTO'
        '404':
          description: Curso ou topico não encontrados
  /cursos/{id1}/topicos/{id2}/respostas:
    get:
      tags:
        - cursos
      summary: Retorna todas as respostas de um topico de um determinado curso
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespostasDTO'
        '404':
          description: Curso, topico não encontrados
    post:
      tags:
        - cursos
      summary: Adiciona uma resposta a um topico de um determinado curso
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespostaFormDTO'
      responses:
        '201':
          description: Cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespostaDTO'
        '404':
          description: Curso ou topico não encontrados
  /cursos/{id1}/topicos/{id2}/respostas/{id3}:
    get:
      tags:
        - cursos
      summary: Retorna uma resposta de um topico de um determinado curso
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
        - name: id3
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespostaDTO'
        '404':
          description: Curso, topico ou resposta não encontrados
    put:
      tags:
        - cursos
      summary: Adiciona uma resposta a um topico de um determinado curso
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
        - name: id3
          in: path
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespostaFormDTO'
      responses:
        '201':
          description: Cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespostaDTO'
        '404':
          description: Curso, topico ou resposta não encontrados
    patch:
      tags:
        - cursos
      summary: Torna uma resposta como solução
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
        - name: id3
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespostaDTO'
        '404':
          description: Curso, topico ou resposta não encontrados
    delete:
      tags:
        - cursos
      summary: Remove uma determinada resposta
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
        - name: id3
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
        '404':
          description: Curso, topico ou resposta não encontrados
...
  /sprints/{id1}/cursos/{id2}:
    post:
      tags:
        - sprints
      summary: Adiciona um curso a uma determinada sprint
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
      responses:
        '201':
          description: Cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SprintDTO'
        '404':
          description: Sprint ou curso não encontrados
    delete:
      tags:
        - sprints
      summary: Remove um determinado curso de uma determinada sprint
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
        '404':
          description: Sprint ou curso não encontrados
...
security:
  - auth: []
components:
  securitySchemes:
    auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CursoDTO:
      type: object
      properties:
        id:
          type: integer
          example: 2
        nome:
          type: string
          example: Spring Boot
        categoria:
          type: string
          example: React
        topicosIds:
          type: array
          items:
            type: string
            example: 'Id: 3 Título: Primeira Sprint Status: FECHADO'
    CursosDTO:
      type: array
      items:
        $ref: '#/components/schemas/CursoDTO'
    CursoFormDTO:
      type: object
      properties:
        nome:
          type: string
          example: Spring Boot
        categoria:
          type: string
          example: React
    RespostaDTO:
      type: object
      properties:
        id:
          type: integer
          example: 1
        autor:
          type: string
          example: João da Silva
        topico:
          type: string
          example: Duvida
        mensagem:
          type: string
          example: Preciso de esclarecimentos sobre o conteúdo da aula um
        dataCriação:
          type: string
          format: date=time
          example: '2021-12-02T00:00:00.000Z'
        solucao:
          type: boolean
          example: true
        arquivo:
          type: string
          example: apostila.pdf
    RespostasDTO:
      type: array
      items:
        $ref: '#/components/schemas/RespostaDTO'
    RespostaFormDTO:
      type: object
      properties:
        IdAutor:
          type: integer
          example: 2
        idTopico:
          type: integer
          example: 2
        mensagem:
          type: string
          example: Resposta do tópico criado
    SprintDTO:
      type: object
      properties:
        id:
          type: integer
          example: 3
        titulo:
          type: string
          example: Primeira Sprint
...
    TopicoDTO:
      type: object
      properties:
        id:
          type: integer
          example: 5
        titulo:
          type: string
          example: Duvida sobre a aula numero dois
        descricao:
          type: string
          example: Detalhamento da duvida em questâo
        dataCriacao:
          type: string
          format: date-time
          example: '2021-10-16T00:00:00.000Z'
        autor:
          type: string
          example: Joao da Silva
        curso:
          type: string
          example: Spring Boot
        status:
          type: string
          example: FECHADO
...
    TopicoFormDTO:
      type: object
      properties:
        titulo:
          type: string
          example: Duvida sobre a aula numero dois
        descricao:
          type: string
          example: Detalhamento da duvida em questão
        cursoId:
          type: integer
          example: 6
        autorId:
          type: integer
          example: 4
...